requirements/base.txt
django
django-extensions
djangorestframework
django-cors-headers
django-environ